% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_vol.R
\name{discharge_vol}
\alias{discharge_vol}
\title{This function determines the discharge volume along different parts of the stream}
\usage{
discharge_vol(depth, width = 2, velocity)
}
\arguments{
\item{depth}{to bottom of the stream bed (feet)}

\item{width}{distance between two points across the stream (feet), default = 2}

\item{velocity}{rate of stream flow (feet per second)}
}
\value{
discharge (cubic feet per second)
}
\description{
This function determines the discharge volume along different parts of the stream
}
