% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\docType{data}
\name{stream_features}
\alias{stream_features}
\title{Generate Stream Features Dataset}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
stream_features
}
\arguments{
\item{nsample}{Number of samples to generate.}
}
\value{
Writes a CSV file containing the generated stream features dataset.
}
\description{
This function runs samples to create a dataset output for use in discharge volume calculations.
}
\examples{
stream_features(nsample=100)
}
\keyword{datasets}
